Refactor MasqMonitor to integrate ReportGenerator for report generation
- Added ReportGenerator as a dependency in MasqMonitor.
- Replaced internal report generation methods with calls to ReportGenerator.
- Simplified domain and URL defanging methods by delegating to ReportGenerator.
- Removed redundant code related to TLP level determination and report generation.
- Updated run_query and save_results methods to utilize ReportGenerator for generating reports.
Resolves Refactor code to move report generation methods into its own file #69
